[PAN-2588] Create P2PNetwork Builder (#1343)
