+++ ethereum/eth/src/test/java/tech/pegasys/pantheon/ethereum/eth/transactions/TestNode.java
@@ -41,7 +41,7 @@ import tech.pegasys.pantheon.ethereum.p2p.api.PeerConnection;
 import tech.pegasys.pantheon.ethereum.p2p.config.DiscoveryConfiguration;
 import tech.pegasys.pantheon.ethereum.p2p.config.NetworkingConfiguration;
 import tech.pegasys.pantheon.ethereum.p2p.config.RlpxConfiguration;
-import tech.pegasys.pantheon.ethereum.p2p.netty.NettyP2PNetwork;
+import tech.pegasys.pantheon.ethereum.p2p.network.DefaultP2PNetwork;
 import tech.pegasys.pantheon.ethereum.p2p.peers.DefaultPeer;
 import tech.pegasys.pantheon.ethereum.p2p.peers.Endpoint;
 import tech.pegasys.pantheon.ethereum.p2p.peers.Peer;
@@ -58,7 +58,6 @@ import java.io.IOException;
 import java.net.InetAddress;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 import java.util.OptionalInt;
 import java.util.concurrent.CompletableFuture;
 
@@ -126,15 +125,14 @@ public class TestNode implements Closeable {
             .protocolManagers(singletonList(ethProtocolManager))
             .network(
                 capabilities ->
-                    new NettyP2PNetwork(
-                        vertx,
-                        this.kp,
-                        networkingConfiguration,
-                        capabilities,
-                        new PeerBlacklist(),
-                        new NoOpMetricsSystem(),
-                        Optional.empty(),
-                        Optional.empty()))
+                    DefaultP2PNetwork.builder()
+                        .vertx(vertx)
+                        .keyPair(this.kp)
+                        .config(networkingConfiguration)
+                        .peerBlacklist(new PeerBlacklist())
+                        .metricsSystem(new NoOpMetricsSystem())
+                        .supportedCapabilities(capabilities)
+                        .build())
             .metricsSystem(new NoOpMetricsSystem())
             .build();
     network = networkRunner.getNetwork();
