+++ ethereum/p2p/src/main/java/tech/pegasys/pantheon/ethereum/p2p/discovery/internal/PeerRequirement.java
@@ -12,8 +12,21 @@
  */
 package tech.pegasys.pantheon.ethereum.p2p.discovery.internal;
 
+import java.util.Collection;
+
 @FunctionalInterface
 public interface PeerRequirement {
 
   boolean hasSufficientPeers();
+
+  static PeerRequirement combine(final Collection<PeerRequirement> peerRequirements) {
+    return () -> {
+      for (PeerRequirement peerRequirement : peerRequirements) {
+        if (!peerRequirement.hasSufficientPeers()) {
+          return false;
+        }
+      }
+      return true;
+    };
+  }
 }
