+++ ethereum/p2p/src/main/java/tech/pegasys/pantheon/ethereum/p2p/network/netty/exceptions/IncompatiblePeerException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 ConsenSys AG.
+ * Copyright 2018 ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -10,11 +10,11 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.ethereum.p2p.netty.exceptions;
+package tech.pegasys.pantheon.ethereum.p2p.network.netty.exceptions;
 
-public class BreachOfProtocolException extends RuntimeException {
+public class IncompatiblePeerException extends RuntimeException {
 
-  public BreachOfProtocolException(final String message) {
+  public IncompatiblePeerException(final String message) {
     super(message);
   }
 }
