+++ ethereum/p2p/src/test/java/tech/pegasys/pantheon/ethereum/p2p/discovery/PeerDiscoveryTestHelper.java
@@ -232,7 +232,6 @@ public class PeerDiscoveryTestHelper {
       return new MockPeerDiscoveryAgent(
           SECP256K1.KeyPair.generate(),
           config,
-          () -> true,
           blacklist,
           whitelist,
           nodePermissioningController,
