+++ ethereum/p2p/src/test/java/tech/pegasys/pantheon/ethereum/p2p/network/netty/testhelpers/SubProtocolMock.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.ethereum.p2p.netty.testhelpers;
+package tech.pegasys.pantheon.ethereum.p2p.network.netty.testhelpers;
 
 import tech.pegasys.pantheon.ethereum.p2p.wire.SubProtocol;
